+++ pgm_layer.py
@@ -2,14 +2,14 @@ from keras.layers import Layer
 from keras.initializers import Initializer
 
 import numpy as np
-from keras import backend as K
+import tensorflow as tf
 
 
 class my_init(Initializer):
     def __init__(self, w):  # w è il vettore dei pesi con cui inizializzare
         self.w = w
 
-    def __call__(self, shape, dtype=np.float32):
+    def __call__(self, shape, dtype=np.float64):
         return self.w
 
     def get_config(self):
@@ -28,18 +28,19 @@ class custom_pgm(Layer):
     def build(self, input_shape):
         self.init = self.add_weight(name='kernel',
                                     shape=(input_shape[1], self.units),
-                                    initializer=my_init(self.initW), trainable=self.trainable)
+                                    initializer=my_init(self.initW), trainable=self.trainable, dtype=np.int64)
         self.slope = self.add_weight(name='kernel',
                                      shape=(input_shape[1], self.units),
-                                     initializer=my_init(self.slope), trainable=self.trainable)
+                                     initializer=my_init(self.slope), trainable=self.trainable, dtype=np.float64)
         self.intercept = self.add_weight(name='kernel',
                                          shape=(input_shape[1], self.units),
-                                         initializer=my_init(self.intercept), trainable=self.trainable)
+                                         initializer=my_init(self.intercept),
+                                         trainable=self.trainable, dtype=np.float64)
         super(custom_pgm, self).build(input_shape)  # Be sure to call this at the end
 
     def call(self, inputs):
         diff = inputs - self.init
-        prod = diff * self.slope
+        prod = tf.cast(diff, np.float64)*self.slope
         prod += self.intercept
         return prod
